update the homepage (with post and announcement feed)
