+++ lib/features/announcement/ui/home_page/add_announcement_bottomsheet.dart
@@ -5,8 +5,6 @@ import 'package:intl/intl.dart';
 import 'package:vocel/common/utils/colors.dart' as constants;
 import 'package:vocel/features/announcement/controller/announcement_list_controller.dart';
 
-
-
 class AddTripBottomSheet extends HookConsumerWidget {
   AddTripBottomSheet({
     super.key,
@@ -35,11 +33,10 @@ class AddTripBottomSheet extends HookConsumerWidget {
           mainAxisSize: MainAxisSize.min,
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
-            const Text("New Vocel Announcement\n",
-              style: TextStyle(
-                  fontWeight: FontWeight.bold,
-                  fontSize: 20
-              ),),
+            const Text(
+              "New Vocel Announcement\n",
+              style: TextStyle(fontWeight: FontWeight.bold, fontSize: 20),
+            ),
             TextFormField(
               decoration: const InputDecoration(
                 hintText: "Announcement Title",
@@ -50,7 +47,8 @@ class AddTripBottomSheet extends HookConsumerWidget {
                 ),
                 focusedBorder: UnderlineInputBorder(
                   borderSide: BorderSide(
-                    color: Color(constants.primaryDarkTeal), // set the focused border color here
+                    color: Color(constants
+                        .primaryDarkTeal), // set the focused border color here
                   ),
                 ),
               ),
@@ -69,7 +67,7 @@ class AddTripBottomSheet extends HookConsumerWidget {
               textInputAction: TextInputAction.next,
             ),
             const SizedBox(
-              height: 20,
+              height: 15,
             ),
             TextFormField(
               keyboardType: TextInputType.name,
@@ -83,13 +81,14 @@ class AddTripBottomSheet extends HookConsumerWidget {
                 ),
                 focusedBorder: UnderlineInputBorder(
                   borderSide: BorderSide(
-                    color: Color(constants.primaryDarkTeal), // set the focused border color here
+                    color: Color(constants
+                        .primaryDarkTeal), // set the focused border color here
                   ),
                 ),
               ),
               autofocus: true,
               autocorrect: false,
-              textInputAction: TextInputAction.next,
+              textInputAction: TextInputAction.done,
               validator: (value) {
                 if (value != null && value.isNotEmpty) {
                   return null;
@@ -98,115 +97,8 @@ class AddTripBottomSheet extends HookConsumerWidget {
                 }
               },
             ),
-            TextFormField(
-              keyboardType: TextInputType.datetime,
-              controller: startDateController,
-              autofocus: true,
-              autocorrect: false,
-              decoration: const InputDecoration(
-                hintText: "Start Date",
-                enabledBorder: UnderlineInputBorder(
-                  borderSide: BorderSide(
-                    color: Color(constants.primaryLightTeal),
-                  ),
-                ),
-                focusedBorder: UnderlineInputBorder(
-                  borderSide: BorderSide(
-                    color: Color(constants.primaryDarkTeal), // set the focused border color here
-                  ),
-                ),
-              ),
-              textInputAction: TextInputAction.next,
-              validator: (value) {
-                if (value != null && value.isNotEmpty) {
-                  return null;
-                } else {
-                  return 'Enter a valid date';
-                }
-              },
-              onTap: () async {
-                DateTime? pickedDate = await showDatePicker(
-                  context: context,
-                  initialDate: DateTime.now(),
-                  firstDate: DateTime(2000),
-                  lastDate: DateTime(2101),
-                  builder: (BuildContext context, Widget? child) {
-                    return Theme(
-                      data: ThemeData.light().copyWith(
-                        colorScheme: const ColorScheme.light(
-                          primary: Color(constants.primaryDarkTeal),
-                          onPrimary: Colors.white,
-                        ),
-                      ),
-                      child: child!,
-                    );
-                  },
-                );
-
-                if (pickedDate != null) {
-                  startDateController.text = DateFormat.yMd().format(pickedDate);
-                }
-                if (pickedDate != null) {
-                  String formattedDate =
-                  DateFormat('yyyy-MM-dd').format(pickedDate);
-                  startDateController.text = formattedDate;
-                } else {}
-              },
-            ),
-            TextFormField(
-              keyboardType: TextInputType.datetime,
-              controller: endDateController,
-              autofocus: true,
-              autocorrect: false,
-              decoration: const InputDecoration(
-                hintText: "End Date",
-                enabledBorder: UnderlineInputBorder(
-                  borderSide: BorderSide(
-                    color: Color(constants.primaryLightTeal),
-                  ),
-                ),
-                focusedBorder: UnderlineInputBorder(
-                  borderSide: BorderSide(
-                    color: Color(constants.primaryDarkTeal), // set the focused border color here
-                  ),
-                ),
-              ),
-              textInputAction: TextInputAction.done,
-              validator: (value) {
-                if (value != null && value.isNotEmpty) {
-                  return null;
-                } else {
-                  return 'Enter a valid date';
-                }
-              },
-              onTap: () async {
-                DateTime? pickedDate = await showDatePicker(
-                  context: context,
-                  initialDate: DateTime.now(),
-                  firstDate: DateTime(2000),
-                  lastDate: DateTime(2101),
-                  builder: (BuildContext context, Widget? child) {
-                    return Theme(
-                      data: ThemeData.light().copyWith(
-                        colorScheme: const ColorScheme.light(
-                          primary: Color(constants.primaryDarkTeal),
-                          onPrimary: Colors.white,
-                        ),
-                      ),
-                      child: child!,
-                    );
-                  },
-                );
-                if (pickedDate != null) {
-                  String formattedDate =
-                  DateFormat('yyyy-MM-dd').format(pickedDate);
-
-                  endDateController.text = formattedDate;
-                }
-              },
-            ),
             const SizedBox(
-              height: 20,
+              height: 15,
             ),
             TextButton(
                 child: const Text(
@@ -220,19 +112,16 @@ class AddTripBottomSheet extends HookConsumerWidget {
                   }
                   if (currentState.validate()) {
                     ref.read(tripsListControllerProvider).add(
-                      name: announcementNameController.text,
-                      description: descriptionController.text,
-                      startDate: startDateController.text,
-                      endDate: endDateController.text,
-                    );
+                          name: announcementNameController.text,
+                          description: descriptionController.text,
+                        );
                     Navigator.of(context).pop();
                   }
                 } //,
-            ),
+                ),
           ],
         ),
       ),
     );
   }
 }
-
